rework interface for accounts
