+++ js/accounts/Accounts.js
@@ -1,0 +1,5 @@
+const list = (data) => {
+  for (const acctData of Object.keys(data?.thirdPartyAccounts || {})) {
+    console.log(acctData)
+  }
+}
