+++ js/accounts/test/Accounts.test.js
@@ -1,0 +1,14 @@
+/* global describe expect test */
+
+import * as accounts from '../Accounts'
+import { loadOrgState } from '../../lib/org-state'
+
+describe('Account', () => {
+
+
+  describe('list', () => {
+    test('lists account names', () => {
+
+    })
+  })
+})
